+++ src/App.tsx
@@ -139,7 +139,7 @@ class App extends React.Component<{}, IState> {
         if (el) {
             el.setAttribute('theme', localStorage.getItem('river.theme.color') || 'light');
             el.setAttribute('font', localStorage.getItem('river.theme.font') || '2');
-            el.setAttribute('bg', localStorage.getItem('river.theme.bg') || '2');
+            el.setAttribute('bg', localStorage.getItem('river.theme.bg') || '15');
             el.setAttribute('bubble', localStorage.getItem('river.theme.bubble') || '4');
             el.setAttribute('direction', localStorage.getItem('river.lang.dir') || 'ltr');
         }